requirements.txt
streamlit
transformers
sentence-transformers
torch
wikipedia
pillow